BackEnd/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.1
asgiref==3.8.1
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==45.0.7
distlib==0.3.9
Django==5.2.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
mailjet-rest==1.5.1
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.11.0
passlib==1.7.4
pillow==11.3.0
platformdirs==4.3.8
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.4
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.1
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
reportlab==4.4.4
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.2
rsa==4.9.1
sentry-sdk==2.33.0
setuptools==80.4.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.46.2
typer==0.16.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.34.2
virtualenv==20.31.2
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
